Make use of the keyring dependency optional in DRE, and build the dre-embedded tool which we ship with the containers without the keyring dependency, to eliminate the need for libdbus1 in the container.
