+++ rs/cli/src/pin/mod.rs
@@ -1,0 +1,3 @@
+pub mod ask;
+#[cfg(feature = "keyring")]
+pub mod keyring;
